Move busy cursor function to new py file
